Don't put components in entrypoint
